+++ src/main/java/at/hannibal2/skyhanni/features/inventory/chocolatefactory/ChocolateFactoryDataLoader.kt
@@ -1,0 +1,358 @@
+package at.hannibal2.skyhanni.features.inventory.chocolatefactory
+
+import at.hannibal2.skyhanni.events.InventoryCloseEvent
+import at.hannibal2.skyhanni.events.InventoryUpdatedEvent
+import at.hannibal2.skyhanni.events.LorenzWorldChangeEvent
+import at.hannibal2.skyhanni.utils.InventoryUtils
+import at.hannibal2.skyhanni.utils.ItemUtils.getLore
+import at.hannibal2.skyhanni.utils.ItemUtils.name
+import at.hannibal2.skyhanni.utils.LorenzUtils.round
+import at.hannibal2.skyhanni.utils.NumberUtil.formatDouble
+import at.hannibal2.skyhanni.utils.NumberUtil.formatInt
+import at.hannibal2.skyhanni.utils.NumberUtil.formatLong
+import at.hannibal2.skyhanni.utils.NumberUtil.romanToDecimal
+import at.hannibal2.skyhanni.utils.SimpleTimeMark
+import at.hannibal2.skyhanni.utils.SoundUtils
+import at.hannibal2.skyhanni.utils.StringUtils.matchFirst
+import at.hannibal2.skyhanni.utils.StringUtils.matchMatcher
+import at.hannibal2.skyhanni.utils.StringUtils.matches
+import at.hannibal2.skyhanni.utils.StringUtils.removeColor
+import at.hannibal2.skyhanni.utils.TimeUtils
+import net.minecraft.item.ItemStack
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent
+
+object ChocolateFactoryDataLoader {
+
+    private val config get() = ChocolateFactoryAPI.config
+    private val profileStorage get() = ChocolateFactoryAPI.profileStorage
+
+    private val chocolatePerSecondPattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "chocolate.persecond",
+        "§6(?<amount>[\\d.,]+) §8per second"
+    )
+    private val chocolateAllTimePattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "chocolate.alltime",
+        "§7All-time Chocolate: §6(?<amount>[\\d,]+)"
+    )
+    private val prestigeLevelPattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "prestige.level",
+        "§6Chocolate Factory (?<prestige>[IVX]+)"
+    )
+    private val chocolateThisPrestigePattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "chocolate.thisprestige",
+        "§7Chocolate this Prestige: §6(?<amount>[\\d,]+)"
+    )
+    private val chocolateForPrestigePattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "chocolate.forprestige",
+        "§7§cRequires (?<amount>\\w+) Chocolate this.*"
+    )
+    private val chocolateMultiplierPattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "chocolate.multiplier",
+        "§7Total Multiplier: §6(?<amount>[\\d.]+)x"
+    )
+    private val leaderboardPlacePattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "leaderboard.place",
+        "§7You are §8#§b(?<position>[\\d,]+)"
+    )
+    private val leaderboardPercentilePattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "leaderboard.percentile",
+        "§7§8You are in the top §.(?<percent>[\\d.]+)%§8 of players!"
+    )
+    private val barnAmountPattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "barn.amount",
+        "§7Your Barn: §.(?<rabbits>\\d+)§7/§.(?<max>\\d+) Rabbits"
+    )
+    private val timeTowerAmountPattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "timetower.amount",
+        "§7Charges: §.(?<uses>\\d+)§7/§a(?<max>\\d+)"
+    )
+    private val timeTowerStatusPattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "timetower.status",
+        "§7Status: §.§l(?<status>INACTIVE|ACTIVE)(?: §f)?(?<acitveTime>\\w*)"
+    )
+    private val timeTowerRechargePattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "timetower.recharge",
+        "§7Next Charge: §a(?<duration>\\w+)"
+    )
+    private val clickMeRabbitPattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "rabbit.clickme",
+        "§e§lCLICK ME!"
+    )
+    private val rabbitAmountPattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "rabbit.amount",
+        "Rabbit \\S+ - \\[(?<amount>\\d+)].*"
+    )
+    private val upgradeTierPattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "upgradetier",
+        ".*\\s(?<tier>[IVXLC]+)"
+    )
+    private val unemployedRabbitPattern by ChocolateFactoryAPI.patternGroup.pattern(
+        "rabbit.unemployed",
+        "Rabbit \\w+ - Unemployed"
+    )
+
+    @SubscribeEvent
+    fun onInventoryUpdated(event: InventoryUpdatedEvent) {
+        if (!ChocolateFactoryAPI.inChocolateFactory) return
+
+        updateInventoryItems(event.inventoryItems)
+    }
+
+    @SubscribeEvent
+    fun onWorldChange(event: LorenzWorldChangeEvent) {
+        clearData()
+    }
+
+    @SubscribeEvent
+    fun onInventoryClose(event: InventoryCloseEvent) {
+        clearData()
+    }
+
+    private fun clearData() {
+        ChocolateFactoryAPI.inChocolateFactory = false
+        ChocolateFactoryAPI.chocolateFactoryPaused = false
+        ChocolateFactoryAPI.factoryUpgrades.clear()
+        ChocolateFactoryAPI.bestAffordableSlot = -1
+        ChocolateFactoryAPI.bestPossibleSlot = -1
+    }
+
+    fun updateInventoryItems(inventory: Map<Int, ItemStack>) {
+        val profileStorage = profileStorage ?: return
+
+        val chocolateItem = InventoryUtils.getItemAtSlotIndex(ChocolateFactoryAPI.infoIndex) ?: return
+        val prestigeItem = InventoryUtils.getItemAtSlotIndex(ChocolateFactoryAPI.prestigeIndex) ?: return
+        val productionInfoItem = InventoryUtils.getItemAtSlotIndex(ChocolateFactoryAPI.productionInfoIndex) ?: return
+        val leaderboardItem = InventoryUtils.getItemAtSlotIndex(ChocolateFactoryAPI.leaderboardIndex) ?: return
+        val barnItem = InventoryUtils.getItemAtSlotIndex(ChocolateFactoryAPI.barnIndex) ?: return
+        val timeTowerItem = InventoryUtils.getItemAtSlotIndex(ChocolateFactoryAPI.timeTowerIndex) ?: return
+
+        processChocolateItem(chocolateItem)
+        processPrestigeItem(prestigeItem)
+        processProductionItem(productionInfoItem)
+        processLeaderboardItem(leaderboardItem)
+        processBarnItem(barnItem)
+        processTimeTowerItem(timeTowerItem)
+
+        ChocolateFactoryAPI.factoryUpgrades.clear()
+
+        profileStorage.rawChocPerSecond =
+            (ChocolateFactoryAPI.chocolatePerSecond / profileStorage.chocolateMultiplier).toInt()
+        profileStorage.lastDataSave = SimpleTimeMark.now().toMillis()
+
+        ChocolateFactoryStats.updateDisplay()
+
+        processInventory(inventory)
+
+        findBestUpgrades()
+    }
+
+    private fun processChocolateItem(item: ItemStack) {
+        val profileStorage = profileStorage ?: return
+
+        ChocolateFactoryAPI.chocolateAmountPattern.matchMatcher(item.name.removeColor()) {
+            profileStorage.currentChocolate = group("amount").formatLong()
+        }
+        for (line in item.getLore()) {
+            chocolatePerSecondPattern.matchMatcher(line) {
+                ChocolateFactoryAPI.chocolatePerSecond = group("amount").formatDouble()
+            }
+            chocolateAllTimePattern.matchMatcher(line) {
+                profileStorage.chocolateAllTime = group("amount").formatLong()
+            }
+        }
+    }
+
+    private fun processPrestigeItem(item: ItemStack) {
+        val profileStorage = profileStorage ?: return
+
+        prestigeLevelPattern.matchMatcher(item.name) {
+            ChocolateFactoryAPI.currentPrestige = group("prestige").romanToDecimal()
+        }
+        var prestigeCost: Long? = null
+        for (line in item.getLore()) {
+            chocolateThisPrestigePattern.matchMatcher(line) {
+                profileStorage.chocolateThisPrestige = group("amount").formatLong()
+            }
+            chocolateForPrestigePattern.matchMatcher(line) {
+                ChocolateFactoryAPI.chocolateForPrestige = group("amount").formatLong()
+                prestigeCost = ChocolateFactoryAPI.chocolateForPrestige
+            }
+        }
+        val prestigeUpgrade = ChocolateFactoryUpgrade(
+            ChocolateFactoryAPI.prestigeIndex,
+            ChocolateFactoryAPI.currentPrestige,
+            prestigeCost,
+            isPrestige = true
+        )
+        ChocolateFactoryAPI.factoryUpgrades.add(prestigeUpgrade)
+    }
+
+    private fun processProductionItem(item: ItemStack) {
+        val profileStorage = profileStorage ?: return
+
+        item.getLore().matchFirst(chocolateMultiplierPattern) {
+            val currentMultiplier = group("amount").formatDouble()
+            profileStorage.chocolateMultiplier = currentMultiplier
+
+            if (ChocolateFactoryTimeTowerManager.timeTowerActive()) {
+                profileStorage.rawChocolateMultiplier = currentMultiplier - profileStorage.timeTowerLevel * 0.1
+            } else {
+                profileStorage.rawChocolateMultiplier = currentMultiplier
+            }
+        }
+    }
+
+    private fun processLeaderboardItem(item: ItemStack) {
+        ChocolateFactoryAPI.leaderboardPosition = null
+        ChocolateFactoryAPI.leaderboardPercentile = null
+
+        for (line in item.getLore()) {
+            leaderboardPlacePattern.matchMatcher(line) {
+                ChocolateFactoryAPI.leaderboardPosition = group("position").formatInt()
+            }
+            leaderboardPercentilePattern.matchMatcher(line) {
+                ChocolateFactoryAPI.leaderboardPercentile = group("percent").formatDouble()
+            }
+        }
+    }
+
+    private fun processBarnItem(item: ItemStack) {
+        val profileStorage = profileStorage ?: return
+
+        item.getLore().matchFirst(barnAmountPattern) {
+            profileStorage.currentRabbits = group("rabbits").formatInt()
+            profileStorage.maxRabbits = group("max").formatInt()
+            ChocolateFactoryBarnManager.trySendBarnFullMessage()
+        }
+    }
+
+    private fun processTimeTowerItem(item: ItemStack) {
+        val profileStorage = profileStorage ?: return
+
+        for (line in item.getLore()) {
+            timeTowerAmountPattern.matchMatcher(line) {
+                profileStorage.currentTimeTowerUses = group("uses").formatInt()
+                profileStorage.maxTimeTowerUses = group("max").formatInt()
+                ChocolateFactoryTimeTowerManager.checkTimeTowerWarning(true)
+            }
+            timeTowerStatusPattern.matchMatcher(line) {
+                val activeTime = group("acitveTime")
+                if (activeTime.isNotEmpty()) {
+                    // todo in future fix this issue with TimeUtils.getDuration
+                    val formattedGroup = activeTime.replace("h", "h ").replace("m", "m ")
+
+                    val activeDuration = TimeUtils.getDuration(formattedGroup)
+                    val activeUntil = SimpleTimeMark.now() + activeDuration
+                    profileStorage.currentTimeTowerEnds = activeUntil.toMillis()
+                }
+            }
+            timeTowerRechargePattern.matchMatcher(line) {
+                // todo in future fix this issue with TimeUtils.getDuration
+                val formattedGroup = group("duration").replace("h", "h ").replace("m", "m ")
+
+                val timeUntilTower = TimeUtils.getDuration(formattedGroup)
+                val nextTimeTower = SimpleTimeMark.now() + timeUntilTower
+                profileStorage.nextTimeTower = nextTimeTower.toMillis()
+            }
+        }
+    }
+
+    private fun processInventory(inventory: Map<Int, ItemStack>) {
+        val profileStorage = profileStorage ?: return
+
+        ChocolateFactoryAPI.clickRabbitSlot = null
+
+        for ((slotIndex, item) in inventory) {
+            if (config.rabbitWarning && clickMeRabbitPattern.matches(item.name)) {
+                SoundUtils.playBeepSound()
+                ChocolateFactoryAPI.clickRabbitSlot = slotIndex
+            }
+
+            if (slotIndex !in ChocolateFactoryAPI.otherUpgradeSlots && slotIndex !in ChocolateFactoryAPI.rabbitSlots) continue
+
+            val itemName = item.name.removeColor()
+            val lore = item.getLore()
+            val upgradeCost = ChocolateFactoryAPI.getChocolateBuyCost(lore)
+
+            val averageChocolate = ChocolateAmount.averageChocPerSecond().round(2)
+            val isMaxed = upgradeCost == null
+
+            var isRabbit = false
+            var level: Int? = null
+            var newAverageChocolate: Double? = null
+
+            when (slotIndex) {
+                in ChocolateFactoryAPI.rabbitSlots -> {
+                    level = rabbitAmountPattern.matchMatcher(itemName) {
+                        group("amount").formatInt()
+                    } ?: run {
+                        unemployedRabbitPattern.matchMatcher(itemName) {
+                            0
+                        }
+                    } ?: continue
+                    isRabbit = true
+
+                    if (isMaxed) {
+                        val rabbitUpgradeItem = ChocolateFactoryUpgrade(slotIndex, level, null, isRabbit = true)
+                        ChocolateFactoryAPI.factoryUpgrades.add(rabbitUpgradeItem)
+                        continue
+                    }
+
+                    val chocolateIncrease = ChocolateFactoryAPI.rabbitSlots[slotIndex] ?: 0
+                    newAverageChocolate = ChocolateAmount.averageChocPerSecond(rawPerSecondIncrease = chocolateIncrease)
+                }
+
+                in ChocolateFactoryAPI.otherUpgradeSlots -> {
+                    level = upgradeTierPattern.matchMatcher(itemName) {
+                        group("tier").romanToDecimal()
+                    } ?: continue
+
+                    if (slotIndex == ChocolateFactoryAPI.timeTowerIndex) profileStorage.timeTowerLevel = level
+
+                    if (isMaxed) {
+                        val otherUpgrade = ChocolateFactoryUpgrade(slotIndex, level, null)
+                        ChocolateFactoryAPI.factoryUpgrades.add(otherUpgrade)
+                        continue
+                    }
+
+                    newAverageChocolate = when (slotIndex) {
+                        ChocolateFactoryAPI.timeTowerIndex -> ChocolateAmount.averageChocPerSecond(
+                            timeTowerLevelIncrease = 1
+                        )
+
+                        ChocolateFactoryAPI.coachRabbitIndex -> ChocolateAmount.averageChocPerSecond(
+                            baseMultiplierIncrease = 0.01
+                        )
+
+                        else -> {
+                            val otherUpgrade = ChocolateFactoryUpgrade(slotIndex, level, upgradeCost)
+                            ChocolateFactoryAPI.factoryUpgrades.add(otherUpgrade)
+                            continue
+                        }
+                    }
+                }
+            }
+            if (level == null || newAverageChocolate == null || upgradeCost == null) continue
+
+            val extra = (newAverageChocolate - averageChocolate).round(2)
+            val effectiveCost = (upgradeCost / extra).round(2)
+
+            val upgrade =
+                ChocolateFactoryUpgrade(slotIndex, level, upgradeCost, extra, effectiveCost, isRabbit = isRabbit)
+            ChocolateFactoryAPI.factoryUpgrades.add(upgrade)
+        }
+    }
+
+    private fun findBestUpgrades() {
+        val profileStorage = profileStorage ?: return
+
+        val notMaxed = ChocolateFactoryAPI.factoryUpgrades.filter { !it.isMaxed }
+
+        val bestUpgrade = notMaxed.minByOrNull { it.effectiveCost ?: Double.MAX_VALUE }
+        profileStorage.bestUpgradeAvailableAt = bestUpgrade?.canAffordAt?.toMillis() ?: 0
+        ChocolateFactoryAPI.bestPossibleSlot = bestUpgrade?.getValidUpgradeIndex() ?: -1
+
+        val affordAbleUpgrade = notMaxed.filter { it.canAfford() }.minByOrNull { it.effectiveCost ?: Double.MAX_VALUE }
+        ChocolateFactoryAPI.bestAffordableSlot = affordAbleUpgrade?.getValidUpgradeIndex() ?: -1
+    }
+}
